Adds a method for returning all objects
This commit adds getAll method and the tests for it. It just returns all of the objects of that table.
A filter functionalitty was not added as it seems, now, not necessary as its not possible to do an improvement on performance given the way we are retrieving objects now (the most simple and non-efficient way).